+++ SpreadsheetUtility/SpreadsheetUtility.Application/UseCases/Session/ListSessionsQuery.cs
@@ -1,6 +1,7 @@
 using MediatR;
 using SpreadsheetUtility.Application.DTOs.Session;
+using SpreadsheetUtility.Application.Ports;
 
 namespace SpreadsheetUtility.Application.UseCases.Session;
 
-public record ListSessionsQuery : IRequest<ListSessionsResponse>;
+public record ListSessionsQuery(CacheBackend cache = CacheBackend.Memory) : IRequest<ListSessionsResponse>;