+++ SpreadsheetUtility/SpreadsheetUtility.Application/UseCases/Session/ListSessionsQueryHandler.cs
@@ -1,4 +1,4 @@
-using MediatR;
+﻿using MediatR;
 using SpreadsheetUtility.Application.DTOs.Session;
 using SpreadsheetUtility.Application.Ports;
 
@@ -6,16 +6,17 @@ namespace SpreadsheetUtility.Application.UseCases.Session;
 
 public class ListSessionsQueryHandler : IRequestHandler<ListSessionsQuery, ListSessionsResponse>
 {
-    private readonly IAuthService _authService;
+    private readonly IAuthServiceFactory _factory;
 
-    public ListSessionsQueryHandler(IAuthService authService)
+    public ListSessionsQueryHandler(IAuthServiceFactory factory)
     {
-        _authService = authService;
+        _factory = factory;
     }
 
     public Task<ListSessionsResponse> Handle(ListSessionsQuery request, CancellationToken cancellationToken)
     {
-        var sessions = _authService.GetAllSessions();
+        var sessionStore = _factory.GetService(request.cache);
+        var sessions = sessionStore.GetAllSessions();
         return Task.FromResult(new ListSessionsResponse(sessions));
     }
 }