+++ SpreadsheetUtility/SpreadsheetUtility.Application/UseCases/Session/UpdateSessionCommand.cs
@@ -1,6 +1,7 @@
 using MediatR;
 using SpreadsheetUtility.Application.DTOs.Session;
+using SpreadsheetUtility.Application.Ports;
 
 namespace SpreadsheetUtility.Application.UseCases.Session;
 
-public record UpdateSessionCommand(string Email, Guid SessionId, string NewValue) : IRequest<UpdateSessionResponse>;
+public record UpdateSessionCommand(string Email, Guid SessionId, string NewValue, CacheBackend cache = CacheBackend.Memory) : IRequest<UpdateSessionResponse>;