+++ SpreadsheetUtility/SpreadsheetUtility.Application/UseCases/Session/UpdateSessionCommandHandler.cs
@@ -1,4 +1,4 @@
-using MediatR;
+﻿using MediatR;
 using SpreadsheetUtility.Application.DTOs.Session;
 using SpreadsheetUtility.Application.Ports;
 
@@ -6,16 +6,17 @@ namespace SpreadsheetUtility.Application.UseCases.Session;
 
 public class UpdateSessionCommandHandler : IRequestHandler<UpdateSessionCommand, UpdateSessionResponse>
 {
-    private readonly IAuthService _authService;
+    private readonly IAuthServiceFactory _factory;
 
-    public UpdateSessionCommandHandler(IAuthService authService)
+    public UpdateSessionCommandHandler(IAuthServiceFactory factory)
     {
-        _authService = authService;
+        _factory = factory;
     }
 
     public Task<UpdateSessionResponse> Handle(UpdateSessionCommand request, CancellationToken cancellationToken)
     {
-        var updatedValue = _authService.UpdateSession(request.Email, request.SessionId, request.NewValue);
+        var sessionStore = _factory.GetService(request.cache);
+        var updatedValue = sessionStore.UpdateSession(request.Email, request.SessionId, request.NewValue);
         return Task.FromResult(new UpdateSessionResponse(updatedValue));
     }
 }